refactor(dashboard): extract the ritmo and insight panels
Dashboard.jsx was 925 lines holding six API calls, eight pieces of state and ten
panels. RitmoCard takes the computed ritmo plus the window size and owns the
heat scale helpers; InsightCard takes the insight and owns the icon heuristic.
Both were self-contained, so the extraction is a move, not a redesign.

754 lines left. The load effect and the row grid stay put: the effect is already
compact and the grid is layout, not logic.

Verified in the browser after a fresh SPA login: both panels render at their
column widths, the heatmap keeps role="img" with its summary, and the grid shows
46 heat cells - 42 days plus the four legend swatches. The same page load also
confirms /recurring/run now fires from AppLayout on a login made inside the SPA.
