+++ frontend/src/pages/Dashboard.jsx
@@ -31,13 +31,15 @@ import { Button } from "@/components/ui/button";
 import { Select } from "@/components/ui/select";
 import NorthStar from "@/components/shared/NorthStar";
 import AiOrb from "@/components/shared/AiOrb";
+import InsightCard from "@/components/dashboard/InsightCard";
+import RitmoCard from "@/components/dashboard/RitmoCard";
 import Money from "@/components/shared/Money";
 import HeroRing from "@/components/shared/HeroRing";
 import { useAuthStore } from "@/store/authStore";
 import { formatDateBR, formatBRL, parseDateOnly } from "@/lib/utils";
 import { colorForCategory } from "@/lib/palette";
 import { emojiForCategory } from "@/lib/categories";
-import { computeRitmo, headroom } from "@/lib/ritmo";
+import { computeRitmo } from "@/lib/ritmo";
 
 // Rótulo curto pt-BR de uma chave ano-mês ("2026-07" → "jul"), em horário local.
 const monthLabel = (ym) => {
@@ -68,16 +70,6 @@ function relativeDay(value) {
   if (diff === 1) return "Ontem";
   if (diff < 7) return `${diff} dias atrás`;
   return formatDateBR(value);
-}
-
-// Ícone contextual dos insights da IA (heurística simples em pt-BR).
-function insightIcon(text) {
-  const t = text.toLowerCase();
-  if (/(caminho certo|parab|bom |ótimo|no azul|guarda|econom|caíram|caiu|reduz)/.test(t))
-    return Check;
-  if (/(crítico|urgente|déficit|acima|estour|exced|negativ|cuidado|risco|falta|imped|ausência|não )/.test(t))
-    return AlertTriangle;
-  return Sparkles;
 }
 
 // Janela do heatmap "Ritmo financeiro" (dias, terminando hoje)
@@ -229,24 +221,6 @@ export default function Dashboard() {
     [streakTx],
   );
 
-  // Intensidade do heatmap: escala sequencial própria (--heat-*), nunca a
-  // paleta categórica do donut — reusá-la aqui faria o painel parecer que
-  // codifica categoria, quando codifica intensidade. 4 = folga total,
-  // 2 = raspou a cota, over = estourou, 0 = dia sem lançamento.
-  function heatLevel(cell) {
-    if (!cell.active) return 0;
-    if (!cell.onPace) return "over";
-    const folga = headroom(cell, ritmo.dailyPace);
-    if (folga > 0.66) return 4;
-    if (folga > 0.33) return 3;
-    return 2;
-  }
-  const heatColor = (level) =>
-    level === "over" ? "rgb(var(--heat-over))" : `rgb(var(--heat-${level}))`;
-  const heatGlow = (level) =>
-    level === "over" || level >= 3
-      ? { boxShadow: `0 0 12px -2px ${heatColor(level)}` }
-      : undefined;
 
   // ── Meta em destaque: a SAVINGS mais próxima de concluir ──
   const featuredGoal = goals
@@ -275,8 +249,6 @@ export default function Dashboard() {
       </div>
     );
   }
-
-  const insightItems = insight?.summary_text?.split("|") || [];
 
   const walletOptions = [
     { value: "all", label: "Todas as carteiras" },
@@ -487,81 +459,7 @@ export default function Dashboard() {
           )}
         </div>
 
-        {/* Ritmo financeiro: dias dentro da cota diária + streak como bônus */}
-        <div className="lg:col-span-5 glass p-6 flex flex-col">
-          <div className="flex items-start justify-between gap-2">
-            <div>
-              <h2 className="font-semibold text-content">Ritmo financeiro</h2>
-              <p className="text-xs text-content-2 mt-0.5">
-                {!ritmo.hasActivity
-                  ? "Registre lançamentos para acompanhar seu ritmo"
-                  : !ritmo.hasPace
-                    ? "Registre uma receita para calcular seu ritmo"
-                    : `${ritmo.onPaceCount} dos últimos ${STREAK_DAYS} dias no seu ritmo`}
-              </p>
-            </div>
-            {/* Só a partir de 3 dias: sequência curta vira cobrança, não prêmio */}
-            {ritmo.hasPace && ritmo.streak >= 3 && (
-              <span className="chip bg-accent/15 text-accent">
-                🔥 {ritmo.streak}
-              </span>
-            )}
-          </div>
-
-          {/* role=img + resumo: o `title` de cada célula é invisível para
-              teclado e ignorado por boa parte dos leitores de tela, então o
-              painel inteiro só existia para quem usa mouse e enxerga. */}
-          <div
-            role="img"
-            aria-label={
-              ritmo.hasPace
-                ? `${ritmo.onPaceCount} dos últimos ${STREAK_DAYS} dias dentro do seu ritmo de gasto diário`
-                : `Sem ritmo calculado nos últimos ${STREAK_DAYS} dias`
-            }
-            className="grid gap-1 mt-4"
-            style={{ gridTemplateColumns: "repeat(14, minmax(0, 1fr))" }}
-          >
-            {ritmo.cells.map((cell, i) => {
-              const level = heatLevel(cell);
-              return (
-                <div
-                  key={cell.key}
-                  title={`${formatDateBR(cell.key)} · ${
-                    cell.active
-                      ? `${formatBRL(cell.spent)} de ${formatBRL(ritmo.dailyPace)}`
-                      : "sem lançamentos"
-                  }`}
-                  style={{
-                    backgroundColor: heatColor(level),
-                    ...heatGlow(level),
-                  }}
-                  className={`heat-cell ${
-                    i === ritmo.cells.length - 1
-                      ? "ring-1 ring-accent ring-offset-1 ring-offset-surface"
-                      : ""
-                  }`}
-                />
-              );
-            })}
-          </div>
-
-          <div className="flex items-center justify-between mt-auto pt-4">
-            <span className="text-[11px] text-content-3">
-              Últimos {STREAK_DAYS} dias
-            </span>
-            <span className="flex items-center gap-1 text-[11px] text-content-3">
-              Menos
-              {[0, 2, 3, 4].map((level) => (
-                <span
-                  key={level}
-                  className="heat-cell w-2.5 h-2.5 shrink-0"
-                  style={{ backgroundColor: heatColor(level) }}
-                />
-              ))}
-              Mais
-            </span>
-          </div>
-        </div>
+        <RitmoCard ritmo={ritmo} dias={STREAK_DAYS} />
 
         {/* Meta em destaque */}
         <div className="lg:col-span-3 relative overflow-hidden glass border-income/25 p-6 flex flex-col">
@@ -734,76 +632,7 @@ export default function Dashboard() {
           )}
         </div>
 
-        {/* Leitura da IA */}
-        <div className="lg:col-span-4 relative overflow-hidden glass border-accent/20 p-6 flex flex-col gap-3">
-          {/* Segundo e último glow do dashboard: presença da IA (ver DESIGN.md) */}
-          <div
-            className="absolute inset-0 pointer-events-none"
-            style={{ background: "var(--glow-accent)" }}
-          />
-          <div className="relative flex items-center gap-3">
-            <AiOrb size={34} />
-            <div>
-              <h2 className="font-semibold text-content">Leitura da IA</h2>
-              <p className="text-[11px] text-accent tracking-wide">
-                resumo do seu comportamento
-              </p>
-            </div>
-          </div>
-          {insightItems.length === 0 ? (
-            <div className="flex-1 flex items-center justify-center text-content-3 text-xs text-center">
-              Adicione transações para gerar sua análise de IA
-            </div>
-          ) : (
-            <div className="flex flex-col gap-2 flex-1">
-              {insightItems.map((item, i) => {
-                const text = item.trim();
-                // 1º insight = destaque; demais ganham chip de ícone contextual
-                if (i === 0) {
-                  return (
-                    <div
-                      key={i}
-                      className="stroke-iris p-3.5 rounded-xl text-[13px] font-semibold text-content leading-relaxed"
-                    >
-                      {text}
-                    </div>
-                  );
-                }
-                const Icon = insightIcon(text);
-                return (
-                  <div
-                    key={i}
-                    className="inset-panel flex items-start gap-3 p-3 text-xs text-content-2 leading-relaxed"
-                  >
-                    <span className="shrink-0 w-9 h-9 rounded-xl bg-accent/[0.12] border border-accent/25 grid place-items-center text-accent">
-                      <Icon size={16} />
-                    </span>
-                    {text}
-                  </div>
-                );
-              })}
-            </div>
-          )}
-
-          {insight?.suggested_action && (
-            <div className="p-3 rounded-xl bg-accent/10 border border-accent/20">
-              <p className="text-[11px] font-semibold text-accent mb-1 uppercase tracking-wider">
-                Sugestão prática
-              </p>
-              <p className="text-xs text-content-2">
-                {insight.suggested_action}
-              </p>
-            </div>
-          )}
-
-          <Button
-            onClick={() => navigate("/ai")}
-            variant="ghost"
-            className="w-full stroke-iris bg-transparent text-accent font-semibold hover:bg-accent/[0.06]"
-          >
-            Conversar com a Norby <ArrowRight size={14} />
-          </Button>
-        </div>
+        <InsightCard insight={insight} />
       </div>
 
       {/* ── Linha 4: gastos por categoria + movimentações recentes ──── */}
